refactor: optimize database schema for type consistency, indexing, and alignment
Type consistency:
- Change Order.UserId from string to Guid to match User.Id PK type
- Change FuelVoucher.AssignedToUserId from string? to Guid?
- Change CreateCheckoutCommand.UserId from string? to Guid?
- Change GetUserPurchasesCommand.UserId from string to Guid
- Change GetUserVouchersCommand.UserId from string to Guid
- Change MarkVoucherAsUsedCommand.UserId from string to Guid
- Change UpdateVoucherCommand.AssignedToUserId from string? to Guid?
- Change BulkActionVouchersCommand.TargetUserId from string? to Guid?
- Change GetSyncCommand.UserId from string to Guid

Remove Purchase table:
- Purchase entity was redundant with Order (same data tracked in orders)
- Deleted entity, configuration, all handlers/commands/queries/controller
- Removed DbSet and DI registrations

Merge Voucher into FuelVoucher:
- Voucher entity overlapped with FuelVoucher (both tracked vouchers)
- Added ExternalId field to FuelVoucher
- Deleted Voucher entity and configuration
- Removed DbSet

Add missing FK indexes:
- contracts(user_id, legal_entity_id, station_id)
- user_contracts(user_id, contract_id)
- voucher_import_errors(import_id)

Add composite indexes for common query patterns:
- orders(user_id, status) and orders(user_id, created_at_utc DESC)
- fuel_vouchers(assigned_to_user_id, status)
- fuel_vouchers(external_id) unique filtered
- verification_codes(phone_number, is_used, expires_at_utc DESC)
- refresh_tokens(user_id, is_revoked)

Fix Lat/Lng from text to double precision for geo queries:
- Station.Lat, Station.Lng: string? -> double?
- StationNode.Lat, StationNode.Lng: string? -> double?

Align Liters type to decimal(10,2) across all entities:
- FuelPackage.Liters: int -> decimal
- Order.Liters: int -> decimal
- PurchaseDto.Liters: int -> decimal
- PackageSuggestionDto.Liters: int -> decimal
- Fix truncation casts in GetPackageSuggestionsQueryHandler

Add UpdatedAtUtc to entities that lacked it:
- Station, StationNode, FuelTypeEntity, FuelPackage, Order, Notification

Update all affected handlers, controllers, DTOs, seed data, and test files

+++ backend/src/FuelFlow.API/Features/Stations/GetPackageSuggestions/GetPackageSuggestionsQueryHandler.cs
@@ -36,7 +36,7 @@ public sealed class GetPackageSuggestionsQueryHandler
             var exists = existingPackages.Any(pkg =>
                 pkg.StationId == station.Id &&
                 pkg.FuelTypeId == fuelType.Id &&
-                pkg.Liters == (int)voucher.Liters);
+                pkg.Liters == voucher.Liters);
             if (exists) continue;
 
             var generatedId = $"{station.Id}-{fuelType.Name.Replace(" ", "").ToLowerInvariant()}-{voucher.Liters}";
@@ -47,7 +47,7 @@ public sealed class GetPackageSuggestionsQueryHandler
                 StationName = station.Name,
                 FuelTypeId = fuelType.Id,
                 FuelName = fuelType.Name,
-                Liters = (int)voucher.Liters
+                Liters = voucher.Liters
             });
         }
 
@@ -62,5 +62,5 @@ public sealed class PackageSuggestionDto
     public string StationName { get; set; } = null!;
     public string FuelTypeId { get; set; } = null!;
     public string FuelName { get; set; } = null!;
-    public int Liters { get; set; }
+    public decimal Liters { get; set; }
 }
